fix update

+++ src/app/DanhSachTrang/Admin/partner/partner.component.ts
@@ -1,3 +1,4 @@
+import { LockPartnerDialogComponent } from './../templates/lock-partner-dialog/lock-partner-dialog.component';
 import { RequestServicesUpdate } from './../../../models/RequestServicesUpdate';
 import { PartnerService } from './../../../services/partner.service';
 import { Partner } from './../../../models/Partner';
@@ -8,6 +9,7 @@ import { PartnerDialogComponent } from '../templates/partner-dialog/partner-dial
 import { identifierModuleUrl } from '@angular/compiler';
 import { RequestServicesUpdateService } from 'src/app/services/request-services-update.service';
 import { PartnerRequestDialogComponent } from '../templates/partner-request-dialog/partner-request-dialog.component';
+import { XoadialogComponent } from '../xoadialog/xoadialog.component';
 @Component({
   selector: 'app-partner',
   templateUrl: './partner.component.html',
@@ -33,7 +35,7 @@ export class PartnerComponent implements OnInit {
 
     this.partnerService.get_AllConfirmPartner().subscribe((data) => {
       let listConfirmPartnerFromServer: Partner[] = data;
-      this.confirmPartners.splice(0,this.confirmPartners.length);
+      this.confirmPartners.splice(0, this.confirmPartners.length);
       for (let partner of listConfirmPartnerFromServer) {
         this.confirmPartners.push(partner);
       }
@@ -41,11 +43,11 @@ export class PartnerComponent implements OnInit {
 
     this.requestServicesUpdateService.get_AllRequest().subscribe((data) => {
       let listRequestFromServer: RequestServicesUpdate[] = data;
-      this.listRequest.splice(0,this.listRequest.length);
+      this.listRequest.splice(0, this.listRequest.length);
       for (let request of listRequestFromServer) {
         this.listRequest.push(request);
       }
-      this.listPartnerByRequest.splice(0,this.listPartnerByRequest.length);
+      this.listPartnerByRequest.splice(0, this.listPartnerByRequest.length);
       for (let request of this.listRequest) {
         for (let partner of this.confirmPartners) {
           if (partner.idDoc == request.partnerid) {
@@ -69,8 +71,8 @@ export class PartnerComponent implements OnInit {
     this.router.navigate(['/admin/partner']);
   }
 
-  ConfirmRequestPartner(partnerSelected: Partner){
-    let reqquestSelected = this.listRequest.find(request => {
+  ConfirmRequestPartner(partnerSelected: Partner) {
+    let reqquestSelected = this.listRequest.find((request) => {
       return request.partnerid == partnerSelected.idDoc;
     });
     partnerSelected.choicesServiceList = reqquestSelected?.listService;
@@ -88,8 +90,8 @@ export class PartnerComponent implements OnInit {
     });
   }
 
-  PartnerRequestDetail(event: any, partnerSelected: Partner){
-    let reqquestSelected = this.listRequest.find(request => {
+  PartnerRequestDetail(event: any, partnerSelected: Partner) {
+    let reqquestSelected = this.listRequest.find((request) => {
       return request.partnerid == partnerSelected.idDoc;
     });
     const dialogRef = this.dialog.open(PartnerRequestDialogComponent, {
@@ -108,23 +110,51 @@ export class PartnerComponent implements OnInit {
     return false;
   }
   OpenAccount(partner: Partner) {
-    partner.accountStatus = 'Đang mở';
-    this.partnerService.update_Partner(partner);
-    this.router.navigate(['/admin/partner']);
+    const dialogRef = this.dialog.open(LockPartnerDialogComponent);
+    dialogRef.afterClosed().subscribe((result: any) => {
+      if (result === true) {
+        partner.accountStatus = 'Đang mở';
+        this.partnerService.update_Partner(partner);
+        this.router.navigate(['/admin/partner']);
+      }
+    });
+
+    // partner.accountStatus = 'Đang mở';
+    // this.partnerService.update_Partner(partner);
+    // this.router.navigate(['/admin/partner']);
   }
   LockAccount(partner: Partner) {
-    partner.accountStatus = 'Đang khoá';
-    this.partnerService.update_Partner(partner);
-    this.router.navigate(['/admin/partner']);
+    const dialogRef = this.dialog.open(LockPartnerDialogComponent);
+    dialogRef.afterClosed().subscribe((result: any) => {
+      if (result === true) {
+        partner.accountStatus = 'Đang khoá';
+        this.partnerService.update_Partner(partner);
+        this.router.navigate(['/admin/partner']);
+      }
+    });
+
+    // partner.accountStatus = 'Đang khoá';
+    // this.partnerService.update_Partner(partner);
+    // this.router.navigate(['/admin/partner']);
   }
   DeletePartner(event: any, partner: Partner) {
-    this.partnerService.delete_Partner(partner);
-    this.router.navigate(['/admin/partner']);
-  }
-  DeleteRequestPartner(event: any, partnerSelected: Partner){
-    let reqquestSelected = this.listRequest.find(request => {
-      return request.partnerid == partnerSelected.idDoc;
+    const dialogRef = this.dialog.open(XoadialogComponent);
+    dialogRef.afterClosed().subscribe((result: any) => {
+      if (result === true) {
+        this.partnerService.delete_Partner(partner);
+        this.router.navigate(['/admin/partner']);
+      }
     });
-    this.requestServicesUpdateService.delete_Request(reqquestSelected!);
+  }
+  DeleteRequestPartner(event: any, partnerSelected: Partner) {
+    const dialogRef = this.dialog.open(XoadialogComponent);
+    dialogRef.afterClosed().subscribe((result: any) => {
+      if (result === true) {
+        let reqquestSelected = this.listRequest.find((request) => {
+          return request.partnerid == partnerSelected.idDoc;
+        });
+        this.requestServicesUpdateService.delete_Request(reqquestSelected!);
+      }
+    });
   }
 }
